refactor: fetch best cycle via hook
