+++ src/components/game/SurvivalTracker.tsx
@@ -1,7 +1,6 @@
 'use client';
 
-import { useEffect, useState } from 'react';
-import { createSupabaseBrowserClient } from '@/lib/supabase/browser';
+import { useBestCycle } from '@/hooks/useBestCycle';
 
 interface SurvivalTrackerProps {
   cycle: number;
@@ -10,44 +9,7 @@ interface SurvivalTrackerProps {
 }
 
 export default function SurvivalTracker({ cycle, unrest, threat }: SurvivalTrackerProps) {
-  const [bestCycle, setBestCycle] = useState<number>(cycle);
-
-  useEffect(() => {
-    let isMounted = true;
-    const loadBest = async () => {
-      let localBest = cycle;
-      try {
-        const stored = localStorage.getItem('ad_best_cycle');
-        if (stored) localBest = Math.max(localBest, parseInt(stored, 10));
-      } catch {}
-
-      let remoteBest = 0;
-      try {
-        const supabase = createSupabaseBrowserClient();
-        const { data } = await supabase
-          .from('game_state')
-          .select('cycle')
-          .order('cycle', { ascending: false })
-          .limit(1);
-        if (data && data.length > 0) remoteBest = data[0].cycle as number;
-      } catch {}
-
-      const best = Math.max(localBest, remoteBest);
-      if (isMounted) {
-        setBestCycle(best);
-      }
-      try { localStorage.setItem('ad_best_cycle', String(best)); } catch {}
-    };
-    loadBest();
-    return () => { isMounted = false; };
-  }, [cycle]);
-
-  useEffect(() => {
-    if (cycle > bestCycle) {
-      setBestCycle(cycle);
-      try { localStorage.setItem('ad_best_cycle', String(cycle)); } catch {}
-    }
-  }, [cycle, bestCycle]);
+  const bestCycle = useBestCycle(cycle);
 
   const bar = (label: string, value: number) => {
     const pct = Math.max(0, Math.min(100, value));
